Finalized Fetch Dashboard data as JSON

+++ Trello.Core/Repositories/Base/IRepository.cs
@@ -9,7 +9,7 @@ namespace Trello.Core.Repositories.Base
     public interface IRepository<T> where T : class
     {
         Task<IReadOnlyList<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(Int64 id);
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(T entity);
